Simplify translation of `=>` to fix bugs.
Previously the left-hand side of `=>` was translated in the same way as a
sequence without `=>`, but that makes it impossible to figure out at
macro-translate time how many arguments are going to be passed to the closure.
The new translation always uses the `UserType` form of each term on the left,
not `RawOutput` and `TupleConcat`, so it never elides terms, even string
constants. Should compile to equally efficient code though.

Closes #3.